requirements.txt
fastapi
uvicorn
websockets
faiss-cpu
sentence-transformers
groq
python-dotenv
streamlit